Menu estadisticas.
Se agrega el menu para el apartado de estadisticas.
