+++ .vscode/settings.json
@@ -9,6 +9,7 @@
         "manipulararchivos.h": "c",
         "stdlib.h": "c",
         "stdio.h": "c",
-        "ctype.h": "c"
+        "ctype.h": "c",
+        "estadisticas.h": "c"
     }
 }
