+++ include/estadisticas.h
@@ -1,0 +1,14 @@
+#include <mysql/mysql.h>
+
+void PrimerEstadistica(MYSQL *conexion);
+
+void SegundaEstadistica(MYSQL *conexion);
+
+void TerceraEstadistica(MYSQL *conexion);
+
+void CuartaEstadistica(MYSQL *conexion);
+
+void QuintaEstadistica(MYSQL *conexion);
+
+void SextaEstadistica(MYSQL *conexion);
+
