+++ src/main.c
@@ -21,8 +21,12 @@
 //./punto_de_venta
 
 
-// Actual.
+// Anterio.
 // gcc -o punto_de_venta src/main.c src/familia.c src/manipularArchivos.c src/producto.c src/cargaStock.c src/db.c src/menu.c src/administrativas.c src/generales.c src/consultaCatalogo.c src/cotizacion.c  -lmysqlclient
+
+
+// Actual.
+// gcc -o punto_de_venta src/main.c src/familia.c src/manipularArchivos.c src/producto.c src/cargaStock.c src/db.c src/menu.c src/administrativas.c src/generales.c src/consultaCatalogo.c src/cotizacion.c src/Estadisticas.c src/facturacion.c  -lmysqlclient
 
 
 int main() {
